requirements.txt
fitz==0.0.1.dev2
Pillow==11.1.0
PySide6==6.6.3.1
PySide6==6.8.2
PySide6_Addons==6.6.3.1
PySide6_Essentials==6.6.3.1
